Update spec config
