+++ spec/rails_helper.rb
@@ -1,32 +1,37 @@
 # This file is copied to spec/ when you run 'rails generate rspec:install'
 ENV['RAILS_ENV'] ||= 'test'
 require File.expand_path('../../config/environment', __FILE__)
+# Prevent database truncation if the environment is production
+abort("The Rails environment is running in production mode!") if Rails.env.production?
+require 'spec_helper'
 require 'rspec/rails'
+# Add additional requires below this line. Rails is not loaded until this point!
 
+# Requires supporting ruby files with custom matchers and macros, etc, in
+# spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are
+# run as spec files by default. This means that files in spec/support that end
+# in _spec.rb will both be required and run as specs, causing the specs to be
+# run twice. It is recommended that you do not name files matching this glob to
+# end with _spec.rb. You can configure this pattern with the --pattern
+# option on the command line or in ~/.rspec, .rspec or `.rspec-local`.
+#
+# The following line is provided for convenience purposes. It has the downside
+# of increasing the boot-up time by auto-requiring all files in the support
+# directory. Alternatively, in the individual `*_spec.rb` files, manually
+# require only the support files necessary.
+#
+Dir[Rails.root.join('spec/support/**/*.rb')].each { |f| require f }
+
+# Checks for pending migration and applies them before tests are run.
+# If you are not using ActiveRecord, you can remove this line.
 ActiveRecord::Migration.maintain_test_schema!
 
 RSpec.configure do |config|
-  config.order                      = 'random'
   config.fixture_path               = "#{::Rails.root}/spec/fixtures"
-  config.global_fixtures            = :all
   config.use_transactional_fixtures = true
+  config.global_fixtures            = :all
 
   config.render_views
+  config.filter_rails_from_backtrace!
   config.infer_spec_type_from_file_location!
-
-  def cookie(name, value)
-    @request.cookies[name.to_s] = CGI::Cookie.new name.to_s, value.to_s
-  end
-
-  def running(&block)
-    lambda &block
-  end
-
-  def login_as(name)
-    @current_user = users(name)
-    @request.session[:user_id] = @current_user.id
-    cookie :user_id, @current_user.id
-    cookie :password_hash, @current_user.password_hash
-  end
-
 end
