+++ spec/support/sugar.rb
@@ -1,0 +1,3 @@
+def running(&block)
+  lambda &block
+end
